+++ pkg/kubeapiserver/options/BUILD
@@ -25,7 +25,6 @@ go_library(
         "//plugin/pkg/admission/certificates/approval:go_default_library",
         "//plugin/pkg/admission/certificates/signing:go_default_library",
         "//plugin/pkg/admission/certificates/subjectrestriction:go_default_library",
-        "//plugin/pkg/admission/defaultingressclass:go_default_library",
         "//plugin/pkg/admission/defaulttolerationseconds:go_default_library",
         "//plugin/pkg/admission/deny:go_default_library",
         "//plugin/pkg/admission/eventratelimit:go_default_library",
@@ -36,6 +35,7 @@ go_library(
         "//plugin/pkg/admission/limitranger:go_default_library",
         "//plugin/pkg/admission/namespace/autoprovision:go_default_library",
         "//plugin/pkg/admission/namespace/exists:go_default_library",
+        "//plugin/pkg/admission/network/defaultingressclass:go_default_library",
         "//plugin/pkg/admission/noderestriction:go_default_library",
         "//plugin/pkg/admission/nodetaint:go_default_library",
         "//plugin/pkg/admission/podnodeselector:go_default_library",
